csc operator/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
rapidfuzz==3.9.1
networkx==3.3
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
pytesseract==0.3.13
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
Pillow==10.4.0
